Fixup nginx SSL
The letsencrypt task will concat the letsencrypt cert with the chain
cert to make a file that nginx can use. Unfortunately because we were
overriding the letsencrypt cert path to be the combined file this file
was growing longer and longer as it was concatenated with itself. This
would cause SSL connections to receive an invalid chain and fail.

Clean this up so that the single certificate file and combined file are
properly seperate.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>