Default to python3 for zuul 3
Setting python_version in vars/vX.yml is too late. These get included in
tasks and so are not present until after meta has run so they are not
actually being picked up.

We could set this in our deployment, but just change the default so that
v3 uses python3 by default.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>