Hacky scripts to get softlayer working
Two scripts to get softlayer boxes to the point where we can actually
use them.

Install python on the image. Newer ubuntu images don't come with python
preinstalled which causes a lot of problems for ansible. We worked
around this in OpenStack by uploading our own images that have python
installed but softlayer makes that more difficult.

Softlayer gives you access to a root account on the boxes which we don't
really want and have an openssh rule to lock down. Add a script to
create a user with the right ssh keys so we can trigger ansible runs.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>