Move flush_handlers out of conditional file
When you have a meta: flush_handlers command it doesn't respect when:
conditions. That means in v3 it'll flush handlers vefore the v3 file has
actually installed them and so will fail. It should succeed on second
run but this is still wrong.

Combine things a bit so flush_handlers is run in the right spot.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>