The log server has to use apache
There's no nginx transition written for the crazy loganalyze stuff that
the log server does. Make it so that the log server always uses apache.

This may not work if opentech-sl takes priority, just going to do it and
see if it needs to be a host_var or something.

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>