Install an ansible fact for bindep_command
We're installing bindep into our base images but to have the zuul
ansible pick it up in a non-standard virtualenv location we need to set
the bindep_command variable. We can set this from our zuul environment
however it really makes more sense to set the fact from the same place
as the command is installed.

This is not really tested. Do it live!

Signed-off-by: Jamie Lennox <jamielennox@gmail.com>